Backend-side/MovieRater/requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
python-decouple==3.3
pytz==2020.1
static3==0.7.0
